Refactor the Key interface
- Moved Key -> Interface::Key

- Interface::Key now includes everything that used to be in Interface::Coder

- Interface::AsymmetricKey and Interface::SymmetricKey are replaced with
  SymmetricKey and AsymmetricKey classes at the top level which descend
  from Interface::Key
