+++ lib/paseto/v3/local.rb
@@ -5,11 +5,9 @@
 module Paseto
   module V3
     # PASETOv3 `local` token interface providing symmetric encryption of tokens.
-    class Local < Key
+    class Local < SymmetricKey
       extend T::Sig
       extend T::Helpers
-
-      include Interface::Symmetric
 
       # Size in bytes of a SHA384 digest
       SHA384_DIGEST_LEN = 48
